Fix #34.
When using nested Maven projects with M2E, it seems M2E does not
properly indicate files belonging to subprojects as filtered from
the parent project. As a result, if we called
IWorkspaceRoot.getFileForLocation(path) to resolve the location
of a resource, it would typically return a IFile referring to the
file through the parent project instead of the expected
subproject. This would then end up opening files in Eclipse
appearing to belong to the "wrong" project (when "Link with
editor" is active) when following a hyperlink.

To work around this, we instead fetch ALL possible IFile
references for the given path and try to pick the best one using
heuristics.